# Make consistent_hash.GetAllReplicas 20 times faster