In benchmarks, this makes distributed FindMissing ~10% faster. GetAllReplicas is also 2.6% of our app CPU, so this will nearly eliminate that.

Benchmark results from `//server/util/consistent_hash:consistent_hash_test`:
```
                                      │    50base     │            50finalnomod2            │
                                      │    sec/op     │   sec/op     vs base                │
GetAllReplicas/CRC32/100_vnodes-24      23560.0n ± 1%   579.5n ± 3%  -97.54% (p=0.000 n=11)
GetAllReplicas/SHA256/10000_vnodes-24   17076.0n ± 3%   759.3n ± 1%  -95.55% (p=0.000 n=11)
geomean                                   20.06µ        663.3n       -96.69%

                                      │   50base   │            50finalnomod2            │
                                      │    B/op    │    B/op     vs base                 │
GetAllReplicas/CRC32/100_vnodes-24      960.0 ± 0%   960.0 ± 0%       ~ (p=1.000 n=11) ¹
GetAllReplicas/SHA256/10000_vnodes-24   960.0 ± 0%   960.0 ± 0%       ~ (p=1.000 n=11) ¹
geomean                                 960.0        960.0       +0.00%

                                      │   50base   │            50finalnomod2            │
                                      │ allocs/op  │ allocs/op   vs base                 │
GetAllReplicas/CRC32/100_vnodes-24      2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=11) ¹
GetAllReplicas/SHA256/10000_vnodes-24   2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=11) ¹
geomean                                 2.000        2.000       +0.00%
```

Benchmark results from `//enterprise/server/test/performance/cache:cache_test`:
```
                               │  cachebase   │           cachefinalnomod           │
                               │    sec/op    │   sec/op     vs base                │
FindMissing/DistDisk10-24        353.8µ ± 12%   293.1µ ± 9%  -17.15% (p=0.000 n=11)
FindMissing/DistDisk100-24       462.3µ ±  4%   417.3µ ± 8%   -9.74% (p=0.000 n=11)
FindMissing/DistDisk1000-24      466.7µ ±  4%   428.9µ ± 3%   -8.10% (p=0.000 n=11)
FindMissing/DistDisk10000-24     480.7µ ±  2%   419.1µ ± 5%  -12.81% (p=0.000 n=11)
FindMissing/DistPebble10-24      459.9µ ±  8%   443.9µ ± 8%        ~ (p=0.088 n=11)
FindMissing/DistPebble100-24     652.7µ ±  1%   599.0µ ± 1%   -8.23% (p=0.000 n=11)
FindMissing/DistPebble1000-24    666.1µ ±  2%   614.5µ ± 3%   -7.74% (p=0.000 n=11)
FindMissing/DistPebble10000-24   653.3µ ±  2%   601.3µ ± 1%   -7.95% (p=0.000 n=11)
geomean                          513.0µ         464.3µ        -9.48%

                               │   cachebase   │           cachefinalnomod            │
                               │     B/op      │     B/op       vs base               │
FindMissing/DistDisk10-24        198.2Ki ±  5%   193.4Ki ± 13%       ~ (p=0.562 n=11)
FindMissing/DistDisk100-24       264.8Ki ±  0%   265.6Ki ±  0%  +0.30% (p=0.035 n=11)
FindMissing/DistDisk1000-24      264.6Ki ±  0%   265.3Ki ±  0%       ~ (p=0.088 n=11)
FindMissing/DistDisk10000-24     264.4Ki ±  0%   264.3Ki ±  0%       ~ (p=0.949 n=11)
FindMissing/DistPebble10-24      314.6Ki ± 10%   335.0Ki ± 12%       ~ (p=0.519 n=11)
FindMissing/DistPebble100-24     462.6Ki ±  0%   462.5Ki ±  0%       ~ (p=0.748 n=11)
FindMissing/DistPebble1000-24    494.0Ki ±  0%   494.1Ki ±  0%       ~ (p=1.000 n=11)
FindMissing/DistPebble10000-24   463.7Ki ±  0%   463.9Ki ±  0%       ~ (p=0.308 n=11)
geomean                          324.3Ki         326.1Ki        +0.55%

                               │  cachebase   │            cachefinalnomod            │
                               │  allocs/op   │  allocs/op    vs base                 │
FindMissing/DistDisk10-24        2.623k ±  5%   2.561k ± 11%       ~ (p=0.573 n=11)
FindMissing/DistDisk100-24       3.590k ±  0%   3.590k ±  0%       ~ (p=0.738 n=11)
FindMissing/DistDisk1000-24      3.590k ±  0%   3.590k ±  0%       ~ (p=0.214 n=11)
FindMissing/DistDisk10000-24     3.590k ±  0%   3.589k ±  0%       ~ (p=0.395 n=11)
FindMissing/DistPebble10-24      5.603k ± 11%   6.003k ± 11%       ~ (p=0.468 n=11)
FindMissing/DistPebble100-24     8.377k ±  0%   8.377k ±  0%       ~ (p=1.000 n=11) ¹
FindMissing/DistPebble1000-24    8.377k ±  0%   8.377k ±  0%       ~ (p=1.000 n=11) ¹
FindMissing/DistPebble10000-24   8.377k ±  0%   8.377k ±  0%       ~ (p=1.000 n=11) ¹
geomean                          5.014k         5.043k        +0.56%
```