@vanja-p 2025-04-14
I added a reference implementation in the test, that is pretty much a simplified version of the previous implementation. I tested both the previous and new implementation against it with hardcoded values, and they return the same thing.